+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * CASE construct (without aggregations).
 * `rand` function added.
 * Maps can now be stored as vertex/edge properties.
+* `collect` aggregation now supports Map collection.
 
 ### Bug Fixes and Other Changes